Extract findQueryType into util-file
